refactor(forms): merge the in/out Headers forms
Last of the eight utility pairs. Both had the same hdrs computed and the same
@input handlers; they differed only in how a row is laid out.

Converged on out/, which is also what DnsServerDrawer already adopted in
16a7e28. Its flex row gives the key and value inputs equal width, where in/'s
grid2 gave the key a full half and made the value share its half with the
delete button, leaving the two columns misaligned. The add button is out/'s
"+ Add" text button rather than an icon-only one.

That is a visible change for the three remaining callers — in/Hysteria2's
masquerade headers, in/Naive and in/OutJson.

components/forms/in is now 18 files, down from 26 at the start of this pass.
